backend/requirements.txt
# Python Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
cryptography==41.0.7

# WebSocket Support
websockets==12.0
python-socketio==5.10.0
eventlet==0.33.3

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
celery==5.3.4
redis==5.0.1
httpx==0.25.2

# Validation
email-validator==2.1.0
phonenumbers==8.13.27

# File validation & security
python-magic==0.4.27

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality (Development)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
